requirements.txt
msal==1.24.1
requests==2.31.0
python-dotenv==1.0.0
office365-rest-python-client==2.5.3
python-docx==1.1.0 